+++ lua/render-markdown/core/conceal.lua
@@ -12,6 +12,7 @@ local Str = require('render-markdown.lib.str')
 ---@field sections render.md.conceal.Section[]
 
 ---@class render.md.Conceal
+---@field private context render.md.Context
 ---@field private buf integer
 ---@field private level integer
 ---@field private computed boolean
@@ -19,13 +20,15 @@ local Str = require('render-markdown.lib.str')
 local Conceal = {}
 Conceal.__index = Conceal
 
+---@param context render.md.Context
 ---@param buf integer
----@param level integer
+---@param win integer
 ---@return render.md.Conceal
-function Conceal.new(buf, level)
+function Conceal.new(context, buf, win)
     local self = setmetatable({}, Conceal)
+    self.context = context
     self.buf = buf
-    self.level = level
+    self.level = Env.win.get(win, 'conceallevel')
     self.computed = false
     self.lines = {}
     return self
@@ -49,17 +52,24 @@ function Conceal:add(row, entry)
     if type(entry) == 'boolean' then
         line.hidden = entry
     else
-        if entry.width <= 0 then
-            return
-        end
         -- If the section is covered by an existing one don't add it
-        for _, section in ipairs(line.sections) do
-            if section.start_col <= entry.start_col and section.end_col >= entry.end_col then
-                return
-            end
+        if entry.width > 0 and not self:has(line, entry) then
+            table.insert(line.sections, entry)
         end
-        table.insert(line.sections, entry)
     end
+end
+
+---@private
+---@param line render.md.conceal.Line
+---@param entry render.md.conceal.Section
+---@return boolean
+function Conceal:has(line, entry)
+    for _, section in ipairs(line.sections) do
+        if section.start_col <= entry.start_col and section.end_col >= entry.end_col then
+            return true
+        end
+    end
+    return false
 end
 
 ---@param width integer
@@ -77,20 +87,18 @@ function Conceal:adjust(width, character)
     end
 end
 
----@param context render.md.Context
 ---@param node render.md.Node
 ---@return boolean
-function Conceal:hidden(context, node)
+function Conceal:hidden(node)
     -- conceal lines metadata require neovim >= 0.11.0 to function
-    return Env.has_11 and self:line(context, node).hidden
+    return Env.has_11 and self:line(node).hidden
 end
 
----@param context render.md.Context
 ---@param node render.md.Node
 ---@return integer
-function Conceal:get(context, node)
+function Conceal:get(node)
     local result = 0
-    for _, section in ipairs(self:line(context, node).sections) do
+    for _, section in ipairs(self:line(node).sections) do
         if node.start_col < section.end_col and node.end_col > section.start_col then
             local amount = self:adjust(section.width, section.character)
             result = result + amount
@@ -100,12 +108,11 @@ function Conceal:get(context, node)
 end
 
 ---@private
----@param context render.md.Context
 ---@param node render.md.Node
-function Conceal:line(context, node)
+function Conceal:line(node)
     if not self.computed then
-        self:compute(context)
         self.computed = true
+        self:compute()
     end
     local line = self.lines[node.start_row]
     if line == nil then
@@ -116,8 +123,7 @@ end
 
 ---Cached row level implementation of vim.treesitter.get_captures_at_pos
 ---@private
----@param context render.md.Context
-function Conceal:compute(context)
+function Conceal:compute()
     if not self:enabled() then
         return
     end
@@ -128,18 +134,17 @@ function Conceal:compute(context)
     if parser == nil then
         return
     end
-    context:parse(parser)
+    self.context:parse(parser)
     parser:for_each_tree(function(tree, language_tree)
-        self:compute_tree(context, language_tree:lang(), tree:root())
+        self:compute_tree(language_tree:lang(), tree:root())
     end)
 end
 
 ---@private
----@param context render.md.Context
 ---@param language string
 ---@param root TSNode
-function Conceal:compute_tree(context, language, root)
-    if not context:overlaps_node(root) then
+function Conceal:compute_tree(language, root)
+    if not self.context:overlaps_node(root) then
         return
     end
     if not vim.tbl_contains({ 'markdown', 'markdown_inline' }, language) then
@@ -149,7 +154,7 @@ function Conceal:compute_tree(context, language, root)
     if query == nil then
         return
     end
-    context:for_each(function(range)
+    self.context:for_each(function(range)
         for id, node, metadata in query:iter_captures(root, self.buf, range.top, range.bottom) do
             if metadata.conceal_lines ~= nil then
                 local node_range = self:node_range(id, node, metadata)