+++ lua/render-markdown/health.lua
@@ -5,7 +5,7 @@ local state = require('render-markdown.state')
 local M = {}
 
 ---@private
-M.version = '8.1.15'
+M.version = '8.1.16'
 
 function M.check()
     M.start('version')